requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Werkzeug==3.0.1
PyJWT==2.10.1
Flask-CORS==6.0.1
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
bleach==6.1.0
python-dotenv==1.0.0
